Create Binary Tree
For the binary tree, pandas was used to help read the csv file in. This is the only use of the library for this tree. After reading in the file, the treeNode class was defined. In this class, I established what the base value of the tree (the root) was and how that value was established. Following this, recursion was used to call upon the data set to determine if the proceeding values were larger or smaller than our root. Determining this, the code calls for the value to be placed in the correct order from left to right ascending values. In order to account for duplicates, I used an if-else statement to tell the tree to essentially ignore any duplicates by filling in values if the spot is empty, and not duplicating if filled. We can see that the number of duplicates are displayed for each value as well(ran in a separate kernel on Anaconda so it would not interfere with the code). The comparisons are displayed and also run in a separate kernel by adding one for every comparison made of each ordering. This way we can see that there are very few comparisons needed for trees, making them highly efficient