Create Bubble Sort
For the bubble sorting, brute force was used to iteratively sort all values. Pandas was again used to start to import the data set as a csv file. While defining our bubble sort function, I established the length of our data set, this would help later for our iteration. Defining i and j, I set the function to run n-i-1 times in the inner loop and n times in the outer loop. This allowed for enough run throughs to sort all data. The data was set so that when the previous value, arr[j] was greater than the next value, arr[j + 1], swapping would occur and force the larger value to move right and the smaller value to move left. This occurred until every value was in the correct spot. To account for duplicates, I used a different Python function called drop_duplicates(inplace=True) which will remove all instances of multiple values in the same space. Essentially redoing the iteration but this time taking into account the number of comparisons, we see that bubble sorting uses many more comparisons than tree sorting. This is highly inefficient to the program speed. As a result, binary trees are generally a better method to save cost of memory usage.