Create Dictionary Linked List
For this assignment, I found Python to be the program I could understand the most. To start, I imported pandas knowing later I would want to import my data file. Like the last program, we start with defining what a Node is considered to be. Next, we want to define our linked list, giving it very specific methods that it can perform including initialization, adding a node(which also inserts it into the linked list (just to skip a step), the deleteNode function to remove a name from the data set upon command, and our printList command to print the data set out. After defining our linked list, I established the main function to show everything our program will do. It imports our data table and begins to start to sort our info into an index, which it does not sort until using the linked list. The index uses a lambda operator which essentially performs the equation in the notes of 26^2*#+26*#+# with the first 3 letters without all of the mess of the equation. It makes the code simpler and is an operation of Python I am just relatively familiar with. It does not change the result. Next I defined the menu to print the options to show the list, delete a node, show a section of our data, or exit out of the menu. When running the code, this is the first thing you will see with the 4 options. Clicking one will follow the images above. Finally, I call the main() function to allow the program to know that this operation is being performed and called.