Create Hash Table
As per all other assignments involving an import, I used pandas to read in the data as a csv file. Next, I began to define the index upon which the code would use to sort the values alphabetically. I used the lambda function to store the first 3 letters of each name. This is very important as there will be a few names later that share the first 3 letters. Following this, I began to create a hash table to sort the data. It directly reads in each name in the file and reads in its index value, assigned previously by our lambda function. If the index doesn’t already exist, it places the index in the hash table and if it does, it is stored at the same value for the moment. Next, a sorting algorithm is used as the dict function creates the hash list of the data and the sorting method calls all of the values in the index and sorts it numerically. Following this I print out the list of all of the names and the index coming from each name. You can see when 2 names share the same first three letters, it prints them side by side. This is what we call a collision. Next I am using a for loop to define what a collision is considered and defining the size of our hash table. Using this I can change the size of the table and print out the number of collisions of each size table. Based on the screenshots above, it is evident that the larger tables result in less collisions. This means that a larger table may result in more unused space, but help with collision count significantly.